+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWEDEPTEJK3XCQHPQ6MSD550"
+        "01KWEMQK6C7V0E1B7ZAT0HAF4J"
     )
     r"""The unique identifier of the data source"""