+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWEDEPTHF9NP0AG5G2AXBC17"
+        "01KWEMQK6G3W54GVGRR49HSWDD"
     )
     r"""The unique identifier of the data source"""