+++ packages/orq-rc/RELEASES.md
@@ -9139,3 +9139,13 @@ Based on:
 - [python v4.11.0-rc.64] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.64] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.64 - packages/orq-rc
+
+## 2026-07-01 10:50:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.65] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.65] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.65 - packages/orq-rc